Split `_base_classes` into `_receiver` and `_sender`
Also move the receiver exceptions from `_exceptions` to `_receiver` and
the sender exceptions from `_exceptions` to `_sender`.

This avoids circular imports.
