+++ src/frequenz/channels/_exceptions.py
@@ -1,16 +1,9 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Exception classes."""
+"""Base exception classes."""
 
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Any, Generic, TypeVar
-
-if TYPE_CHECKING:
-    from . import _base_classes
-
-_T = TypeVar("_T")
+from typing import Any
 
 
 class Error(RuntimeError):
@@ -56,54 +49,3 @@ class ChannelClosedError(ChannelError):
             channel: A reference to the channel that was closed.
         """
         super().__init__(f"Channel {channel} was closed", channel)
-
-
-class SenderError(Error, Generic[_T]):
-    """An error produced in a [Sender][frequenz.channels.Sender].
-
-    All exceptions generated by senders inherit from this exception.
-    """
-
-    def __init__(self, message: str, sender: _base_classes.Sender[_T]):
-        """Create an instance.
-
-        Args:
-            message: An error message.
-            sender: The [Sender][frequenz.channels.Sender] where the error
-                happened.
-        """
-        super().__init__(message)
-        self.sender: _base_classes.Sender[_T] = sender
-        """The sender where the error happened."""
-
-
-class ReceiverError(Error, Generic[_T]):
-    """An error produced in a [Receiver][frequenz.channels.Receiver].
-
-    All exceptions generated by receivers inherit from this exception.
-    """
-
-    def __init__(self, message: str, receiver: _base_classes.Receiver[_T]):
-        """Create an instance.
-
-        Args:
-            message: An error message.
-            receiver: The [Receiver][frequenz.channels.Receiver] where the
-                error happened.
-        """
-        super().__init__(message)
-        self.receiver: _base_classes.Receiver[_T] = receiver
-        """The receiver where the error happened."""
-
-
-class ReceiverStoppedError(ReceiverError[_T]):
-    """The [Receiver][frequenz.channels.Receiver] stopped producing messages."""
-
-    def __init__(self, receiver: _base_classes.Receiver[_T]):
-        """Create an instance.
-
-        Args:
-            receiver: The [Receiver][frequenz.channels.Receiver] where the
-                error happened.
-        """
-        super().__init__(f"Receiver {receiver} was stopped", receiver)
