+++ src/frequenz/channels/_receiver.py
@@ -1,33 +1,18 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Base classes for Channel Sender and Receiver."""
+"""Channel receiver and associated exceptions."""
 
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections.abc import Callable
-from typing import Generic, TypeVar
+from typing import Generic, Self, TypeVar
 
-from ._exceptions import ReceiverStoppedError
+from ._exceptions import Error
 
 _T = TypeVar("_T")
 _U = TypeVar("_U")
-
-
-class Sender(ABC, Generic[_T]):
-    """A channel Sender."""
-
-    @abstractmethod
-    async def send(self, msg: _T) -> None:
-        """Send a message to the channel.
-
-        Args:
-            msg: The message to be sent.
-
-        Raises:
-            SenderError: if there was an error sending the message.
-        """
 
 
 class Receiver(ABC, Generic[_T]):
@@ -76,7 +61,7 @@ class Receiver(ABC, Generic[_T]):
             ReceiverError: if there is some problem with the receiver.
         """
 
-    def __aiter__(self) -> Receiver[_T]:
+    def __aiter__(self) -> Self:
         """Initialize the async iterator over received values.
 
         Returns:
@@ -121,6 +106,38 @@ class Receiver(ABC, Generic[_T]):
             A `Receiver` to read results of the given function from.
         """
         return _Map(self, call)
+
+
+class ReceiverError(Error, Generic[_T]):
+    """An error produced in a [Receiver][frequenz.channels.Receiver].
+
+    All exceptions generated by receivers inherit from this exception.
+    """
+
+    def __init__(self, message: str, receiver: Receiver[_T]):
+        """Create an instance.
+
+        Args:
+            message: An error message.
+            receiver: The [Receiver][frequenz.channels.Receiver] where the
+                error happened.
+        """
+        super().__init__(message)
+        self.receiver: Receiver[_T] = receiver
+        """The receiver where the error happened."""
+
+
+class ReceiverStoppedError(ReceiverError[_T]):
+    """The [Receiver][frequenz.channels.Receiver] stopped producing messages."""
+
+    def __init__(self, receiver: Receiver[_T]):
+        """Create an instance.
+
+        Args:
+            receiver: The [Receiver][frequenz.channels.Receiver] where the
+                error happened.
+        """
+        super().__init__(f"Receiver {receiver} was stopped", receiver)
 
 
 class _Map(Receiver[_U], Generic[_T, _U]):
