+++ src/frequenz/channels/util/_event.py
@@ -6,10 +6,10 @@
 
 import asyncio as _asyncio
 
-from frequenz.channels import _base_classes, _exceptions
+from frequenz.channels import _receiver
 
 
-class Event(_base_classes.Receiver[None]):
+class Event(_receiver.Receiver[None]):
     """A receiver that can be made ready through an event.
 
     The receiver (the [`ready()`][frequenz.channels.util.Event.ready] method) will wait
@@ -134,7 +134,7 @@ class Event(_base_classes.Receiver[None]):
             ReceiverStoppedError: If this receiver is stopped.
         """
         if not self._is_set and self._is_stopped:
-            raise _exceptions.ReceiverStoppedError(self)
+            raise _receiver.ReceiverStoppedError(self)
 
         assert self._is_set, "calls to `consume()` must be follow a call to `ready()`"
